// frontend/src/types/proto-es/v1/audit_log_service_pb.d.ts
export declare type AuditLogStatus = Message<"metaxisdata.v1.AuditLogStatus"> & {
  /**
   * @generated from field: int32 code = 1;
   */
  code: number;

  /**
   * @generated from field: string message = 2;
   */
  message: string;
};

/**
 * Describes the message metaxisdata.v1.AuditLogStatus.
 * Use `create(AuditLogStatusSchema)` to create a new message.
 */
export declare const AuditLogStatusSchema: GenMessage<AuditLogStatus>;

/**
 * @generated from message metaxisdata.v1.AuditRequestMetadata
 */
export declare type AuditRequestMetadata = Message<"metaxisdata.v1.AuditRequestMetadata"> & {
  /**
   * @generated from field: string ip = 1;
   */
  ip: string;

  /**
   * @generated from field: string user_agent = 2;
   */
  userAgent: string;
};

/**
 * Describes the message metaxisdata.v1.AuditRequestMetadata.
 * Use `create(AuditRequestMetadataSchema)` to create a new message.
 */
export declare const AuditRequestMetadataSchema: GenMessage<AuditRequestMetadata>;

/**
 * @generated from message metaxisdata.v1.AuditLog
 */
export declare type AuditLog = Message<"metaxisdata.v1.AuditLog"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: google.protobuf.Timestamp create_time = 2;
   */
  createTime?: Timestamp;

  /**
   * @generated from field: string parent = 3;
   */
  parent: string;

  /**
   * @generated from field: string method = 4;
   */
  method: string;

  /**
   * @generated from field: string resource = 5;
   */
  resource: string;

  /**
   * @generated from field: string user = 6;
   */
  user: string;

  /**
   * @generated from field: metaxisdata.v1.AuditLogSeverity severity = 7;
   */
  severity: AuditLogSeverity;

  /**
   * @generated from field: google.protobuf.Struct request = 8;
   */
  request?: JsonObject;

  /**
   * @generated from field: google.protobuf.Struct response = 9;
   */
  response?: JsonObject;

  /**
   * @generated from field: metaxisdata.v1.AuditLogStatus status = 10;
   */
  status?: AuditLogStatus;

  /**
   * @generated from field: int64 latency_ms = 11;
   */
  latencyMs: bigint;

  /**
   * @generated from field: google.protobuf.Struct service_data = 12;
   */
  serviceData?: JsonObject;

  /**
   * @generated from field: metaxisdata.v1.AuditRequestMetadata request_metadata = 13;
   */
  requestMetadata?: AuditRequestMetadata;
};
// ...
export enum AuditLogSeverity {
  /**
   * @generated from enum value: AUDIT_LOG_SEVERITY_UNSPECIFIED = 0;
   */
  AUDIT_LOG_SEVERITY_UNSPECIFIED = 0,

  /**
   * @generated from enum value: INFO = 1;
   */
  INFO = 1,

  /**
   * @generated from enum value: WARNING = 2;
   */
  WARNING = 2,

  /**
   * @generated from enum value: ERROR = 3;
   */
  ERROR = 3,
}